+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.14.0](https://github.com/kittysaemi/KittyWallet/compare/v1.13.0...v1.14.0) (2026-06-20)
+
+
+### Features
+
+* **ui:** add icon snapshot fallback ([6ecb4e4](https://github.com/kittysaemi/KittyWallet/commit/6ecb4e4cc14f5a6a4452305ba3db4a21577111d2))
+
+# [1.13.0](https://github.com/kittysaemi/KittyWallet/compare/v1.12.0...v1.13.0) (2026-06-20)
+
+
+### Features
+
+* **api:** add icon snapshot storage ([549d319](https://github.com/kittysaemi/KittyWallet/commit/549d31980221a20a7ceea222d3fb052b77df2e1e))
+
 # [1.12.0](https://github.com/kittysaemi/KittyWallet/compare/v1.11.0...v1.12.0) (2026-06-20)